+++ src/Datasets/Dataset.php
@@ -6,7 +6,7 @@ use Rubix\ML\Report;
 use Rubix\ML\DataType;
 use Rubix\ML\Helpers\Stats;
 use Rubix\ML\Extractors\Writable;
-use Rubix\ML\Transformers\Bidirectional;
+use Rubix\ML\Transformers\Reversible;
 use Rubix\ML\Transformers\Stateful;
 use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Kernels\Distance\Distance;
@@ -292,13 +292,13 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, Countable
     }
 
     /**
-     * Reverse a transformation to the dataset.
+     * Reverse a transformation that was applied to the dataset.
      *
-     * @param \Rubix\ML\Transformers\Bidirectional $transformer
+     * @param \Rubix\ML\Transformers\Reversible $transformer
      * @throws \Rubix\ML\Exceptions\RuntimeException
      * @return static
      */
-    public function reverse(Bidirectional $transformer) : self
+    public function reverseApply(Reversible $transformer) : self
     {
         if ($transformer instanceof Stateful) {
             if (!$transformer->fitted()) {