+++ tests/Transformers/MaxAbsoluteScalerTest.php
@@ -2,10 +2,9 @@
 
 namespace Rubix\ML\Tests\Transformers;
 
-use Rubix\ML\Datasets\Unlabeled;
-use Rubix\ML\Transformers\Bidirectional;
 use Rubix\ML\Transformers\Elastic;
 use Rubix\ML\Transformers\Stateful;
+use Rubix\ML\Transformers\Reversible;
 use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Transformers\MaxAbsoluteScaler;
@@ -47,13 +46,13 @@ class MaxAbsoluteScalerTest extends TestCase
         $this->assertInstanceOf(Transformer::class, $this->transformer);
         $this->assertInstanceOf(Stateful::class, $this->transformer);
         $this->assertInstanceOf(Elastic::class, $this->transformer);
-        $this->assertInstanceOf(Bidirectional::class, $this->transformer);
+        $this->assertInstanceOf(Reversible::class, $this->transformer);
     }
 
     /**
      * @test
      */
-    public function fitUpdateTransform() : void
+    public function fitUpdateTransformReverse() : void
     {
         $this->transformer->fit($this->generator->generate(30));
 
@@ -66,15 +65,23 @@ class MaxAbsoluteScalerTest extends TestCase
         $this->assertIsArray($maxabs);
         $this->assertCount(3, $maxabs);
 
-        $sample = $this->generator->generate(1)
-            ->apply($this->transformer)
-            ->sample(0);
+        $dataset = $this->generator->generate(1);
+
+        $original = $dataset->sample(0);
+
+        $dataset->apply($this->transformer);
+
+        $sample = $dataset->sample(0);
 
         $this->assertCount(3, $sample);
 
         $this->assertEqualsWithDelta(0, $sample[0], 1);
         $this->assertEqualsWithDelta(0, $sample[1], 1);
         $this->assertEqualsWithDelta(0, $sample[2], 1);
+
+        $dataset->reverseApply($this->transformer);
+
+        $this->assertEquals($original, $dataset->sample(0));
     }
 
     /**
@@ -87,37 +94,6 @@ class MaxAbsoluteScalerTest extends TestCase
         $samples = $this->generator->generate(1)->samples();
 
         $this->transformer->transform($samples);
-    }
-
-    /**
-     * @test
-     */
-    public function reverseTransform() : void
-    {
-        $this->transformer->fit($this->generator->generate(30));
-        $samples = $expected = $this->generator->generate(1)->samples();
-
-        $this->transformer->transform($samples);
-        $this->assertNotEquals($expected, $samples);
-
-        $this->transformer->reverseTransform($samples);
-        $this->assertEquals($expected, $samples);
-    }
-
-    /**
-     * @test
-     */
-    public function reverseTransformWithReverseApply() : void
-    {
-        $this->transformer->fit($this->generator->generate(30));
-        $samples = $this->generator->generate(1)->samples();
-        $dataset = Unlabeled::fromIterator($samples);
-
-        $dataset->apply($this->transformer);
-        $this->assertNotEquals($dataset->samples(), $samples);
-
-        $dataset->reverse($this->transformer);
-        $this->assertEquals($dataset->samples(), $samples);
     }
 
     /**